refactor generate prompt
