+++ src/utils/openai.ts
@@ -1,4 +1,5 @@
 import { encode } from "@nem035/gpt-3-encoder";
+import { Schema, Table } from "@/types";
 
 // openAIApiKey is the API key for OpenAI API.
 export const openAIApiKey = process.env.OPENAI_API_KEY;
@@ -9,3 +10,39 @@ export const openAIApiEndpoint = process.env.OPENAI_API_ENDPOINT || "https://api
 export const countTextTokens = (text: string) => {
   return encode(text).length;
 };
+
+export function generateDbPromptFromContext(
+  promptGenerator: (input: string | undefined) => string,
+  schemaList: any,
+  selectedSchemaName: string,
+  selectedTablesName: string[],
+  maxToken: number,
+  userPrompt?: string
+): string {
+  let schema = "";
+  // userPrompt is message that user want to send to bot. When cat prompt in drawer, userPrompt is undefined.
+  let tokens = countTextTokens(userPrompt || "");
+
+  // Empty table name(such as []) denote all table. [] and `undefined` both are false in `if`
+  const tableList: string[] = [];
+  const selectedSchema = schemaList.find((schema: Schema) => schema.name == (selectedSchemaName || ""));
+  if (selectedTablesName) {
+    selectedTablesName.forEach((tableName: string) => {
+      const table = selectedSchema?.tables.find((table: Table) => table.name == tableName);
+      tableList.push(table!.structure);
+    });
+  } else {
+    for (const table of selectedSchema?.tables || []) {
+      tableList.push(table!.structure);
+    }
+  }
+  if (tableList) {
+    for (const table of tableList) {
+      if (tokens < maxToken / 2) {
+        tokens += countTextTokens(table);
+        schema += table;
+      }
+    }
+  }
+  return promptGenerator(schema);
+}
